pyproject.toml
[tool.poetry]
name = "diplom-for-api"
version = "0.1.0"
description = ""
authors = ["widowblack11 <oksana.prokopenko16@gmail.com>"]
readme = "README.md"
packages = [{include = "diplom_for_api"}]

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.2.2"
requests = "^2.28.2"
voluptuous = "^0.13.1"
pytest-voluptuous = "^1.2.0"
selene = "2.0.0b16"
allure-pytest = "^2.13.1"
python-dotenv = "^1.0.0"
curlify = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
